+++ Tests/PowerSyncTests/SyncTests.swift
@@ -740,7 +740,7 @@ class InMemorySyncIntegrationTests {
     }
 
     @Test func subscriptionsUpdateWhileOffline() async throws {
-        let db = openDatabase(PlatformHttpClient.shared)
+        let db = openDatabase(PlatformHttpClient(session: .shared))
         // Make sure the database is initialized
         try await db.readLock { _ in }
         var statusUpdates = db.currentStatus.asFlow().makeAsyncIterator()
@@ -823,7 +823,7 @@ private func openDatabase(_ client: any HttpClient, schema: Schema = defaultSche
         activeInstanceStore: DatabaseGroupCollection(),
         logger: logger,
         pool: AsyncConnectionPool(location: .inMemory, logger: DefaultLogger()),
-        httpClient: client,
+        customHttpClient: client,
         schema: schema,
     )
 }