+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Fix `SyncStatus.asFlow()` emitting the same object, causing SwiftUI to miss updates.
 * Add the `ObservableSyncStatus` utility, which can be used to track Sync Status updates through an `@Observable` class.
 * Clear `SyncStatus.uploadError` after a successful upload.
+* Allow passing a custom `URLSession` in `SyncClientConfiguration`.
 
 ## 1.14.4